requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
jupyterlab~=4.0.6
requests~=2.31.0
djangorestframework-jwt~=
drf-yasg~=
djangorestframework-simplejwt~=